splash screen
Splash Screen added, permission requests
