+++ android/app/build.gradle
@@ -36,7 +36,7 @@ def keystoreProperties = new Properties()
 
 android {
     namespace "com.example.dashboardapp"
-    compileSdkVersion flutter.compileSdkVersion
+    compileSdkVersion 33
     ndkVersion flutter.ndkVersion
 
     compileOptions {
@@ -90,4 +90,5 @@ flutter {
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.multidex:multidex:2.0.1'
+    
 }
