+++ android/app/src/main/res/drawable-night/background.png
