+++ lib/main.dart
@@ -1,19 +1,24 @@
-import 'dart:convert';
-import 'package:dashboardapp/adminscreen.dart';
-import 'package:http/http.dart' as http;
-import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:dashboardapp/loginscreen.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
-import 'userscreen.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:get/get.dart';
+import 'package:permission_handler/permission_handler.dart';
+import 'package:lottie/lottie.dart';
+import 'package:animated_splash_screen/animated_splash_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
+  if (await Permission.notification.isDenied ||
+      await Permission.storage.isDenied) {
+    // Either the permission was already granted before or the user just granted it.
+    Map<Permission, PermissionStatus> statuses = await [
+      Permission.notification,
+      Permission.storage,
+    ].request();
+  }
   runApp(const LoginApp());
 }
 
@@ -27,125 +32,11 @@ class LoginApp extends StatelessWidget {
         useMaterial3: true,
       ),
       debugShowCheckedModeBanner: false,
-      //home: const adminDashBoard(),
-      home: const LoginScreen(),
-    );
-  }
-}
-
-class LoginScreen extends StatefulWidget {
-  const LoginScreen({super.key});
-
-  @override
-  _LoginScreenState createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-  var registrationToken;
-  late Map<String, dynamic> tokenData;
-  final TextEditingController emailController = TextEditingController();
-  var credential;
-  final TextEditingController passwordController = TextEditingController();
-  var headers = {'Content-Type': 'application/json'};
-
-  Future<void> registerDevice() async {
-    registrationToken = await FirebaseMessaging.instance.getToken();
-    tokenData = {"token": registrationToken};
-    var request = http.Request('POST',
-        Uri.parse('https://mobilon-backend.onrender.com/registerDevice'));
-    request.body = json.encode(tokenData);
-    request.headers.addAll(headers);
-
-    http.StreamedResponse response = await request.send();
-
-    if (response.statusCode == 200) {
-      print(await response.stream.bytesToString());
-    } else {
-      print(response.reasonPhrase);
-    }
-  }
-
-  void _handleLogin() async {
-    String email = emailController.text;
-    String password = passwordController.text;
-
-    try {
-      credential = await FirebaseAuth.instance
-          .signInWithEmailAndPassword(email: email, password: password);
-    } on FirebaseAuthException catch (e) {
-      if (e.code == 'user-not-found') {
-        print('No user found for that email.');
-      } else if (e.code == 'wrong-password') {
-        print('Wrong password provided for that user.');
-      }
-    }
-    final user = credential.user;
-    if (user == null) {
-      print('User is currently signed out!');
-    } else {
-      // Register the Device
-      await registerDevice();
-      print("Done");
-
-      if (password.contains("core")) {
-        // ignore: use_build_context_synchronously
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const userDashBoard()),
-        );
-      } else if (password.contains("head")) {
-        // ignore: use_build_context_synchronously
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const adminDashBoard()),
-        );
-      } else {
-        // ignore: use_build_context_synchronously
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text("Invalid email or password."),
-            duration: Duration(seconds: 2),
-          ),
-        );
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Login'),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: SingleChildScrollView(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Center(child: Image.asset('assets/Mobilon.png')),
-              TextField(
-                controller: emailController,
-                decoration: const InputDecoration(
-                  labelText: 'Email',
-                ),
-              ),
-              const SizedBox(height: 20.0),
-              TextField(
-                controller: passwordController,
-                decoration: const InputDecoration(
-                  labelText: 'Password',
-                ),
-                obscureText: true,
-              ),
-              const SizedBox(height: 20.0),
-              ElevatedButton(
-                onPressed: _handleLogin,
-                child: const Text('Login'),
-              ),
-            ],
-          ),
-        ),
+      home: AnimatedSplashScreen(
+        splash: 'assets/Mobilonv1.png',
+        splashIconSize: double.infinity,
+        nextScreen: LoginScreen(),
+        splashTransition: SplashTransition.scaleTransition,
       ),
     );
   }
