+++ pubspec.yaml
@@ -42,12 +42,16 @@ dependencies:
   firebase_core: ^2.15.1
   firebase_auth: ^4.8.0
   multi_dropdown: ^2.0.1
-  cloud_firestore: ^4.9.0
+  cloud_firestore: ^4.12.2
   loading_animation_widget: ^1.2.0+4
   get: ^4.6.5
   webview_flutter: ^4.2.4
-  firebase_messaging: ^14.6.7
+  firebase_messaging: ^14.7.3
   http: ^1.1.0
+  permission_handler: ^11.0.1
+  lottie: ^2.7.0
+  flutter_native_splash: ^2.3.5
+  animated_splash_screen: ^1.3.0
 
 dev_dependencies:
   flutter_test:
