+++ windows/flutter/generated_plugins.cmake
@@ -3,7 +3,10 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  cloud_firestore
+  firebase_auth
   firebase_core
+  permission_handler_windows
   url_launcher_windows
   window_to_front
 )
